Starting to work on refresh function
