+++ peloton_metrics/dal/postgres/tracked_users_dao.py
@@ -5,13 +5,17 @@ from peloton_metrics.dal.postgres.base_dao import BaseDao
 from peloton_metrics.dal.postgres.helpers.json_helper import JsonHelper
 from peloton_metrics.dal.postgres.helpers.sql_type_constants import SqlType
 from peloton_metrics.models.user import User
+from peloton_metrics.dal.postgres.helpers.decorators import select_as_dictionaries
 
 
 class TrackedUsersDao(BaseDao):
 
+    @select_as_dictionaries
     def fetch_tracked_users(self):
         sql = f"""
-            SELECT user_id FROM public.tracked_users;
+            SELECT
+                user_id AS user_id
+            FROM public.tracked_users;
         """
         return self.execute_sql(sql)
