+++ peloton_metrics/run/run_workout_refresh.py
@@ -1,3 +1,4 @@
 from peloton_metrics.processors.refresh_workout_processor import RefreshWorkoutProcessor
 
+# TODO: Change this to use click library to run from CLI.
 RefreshWorkoutProcessor().run()
